pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "sqlparse>=0.5.0",
    "python-dotenv>=1.0.0",
]
